New type of connection pool using arrays.
